requirements.txt
flask
tensorflow
numpy
pillow
requests
flask-login
flask-sqlalchemy
openai
google-generativeai
python-dotenv
